+++ CHANGELOG.md
@@ -11,9 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `InputChips` and `SelectMulti` chip selection and copy-pasting
 - `Select` now supports `showCreate` and `formatCreateLabel` that were previously only supported in `SelectMulti`
+- `GoogleFontsLoader`
+- `ComponentsProvider` now supports `loadGoogleFonts` which leverages `GoogleFontsLoader`
 
 ### Changed
 
+- `theme.fonts.*` updates
+  - `body` added (now default for most components)
+  - `brand` is now used just for `Button*`, `Heading` & `Tabs`
+  - Default values changed (now `body='Roboto'`, `brand='Red Hat Display'`, `code='Roboto Mono'`)
+- `theme.fontWeights.extraBold` & `theme.fontWeights.light` removed
 - `Tree` child `AccordionDisclosure` now receives font-weight value from styled-components selector
 
 ### Fixed
@@ -22,10 +29,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SelectMulti` failing to appropriately show "No options" when `showCreate` is used
 - `Select` overwriting search value with the current option value if the option's value and label are different
 - Allow `Icon` colors to be modified when used inside a `MenuItem` detail prop.
+- Miscellaneous Storybook fixes to silence console warnings
+
 
 ###
 
 - `Tooltip` & `Popover` no longer support (optional) arrow indicator
+
+### Removed
+
+- `theme.colors.palette` is no longer available (`palette` _can_ be import from `@looker/design-tokens` but this is a legacy behavior and is not encouraged)
+- `prismTheme` is no longer published
 
 ## [0.9.13] - 2020-08-24